time-series: fix RegularTimeSeriesIndex::getPointCount for huge periods

+++ time-series/time-series-api/src/main/java/com/powsybl/timeseries/RegularTimeSeriesIndex.java
@@ -112,7 +112,7 @@ public class RegularTimeSeriesIndex extends AbstractTimeSeriesIndex {
 
     @Override
     public int getPointCount() {
-        return Math.round(((float) (endTime - startTime)) / spacing) + 1;
+        return (int) Math.round(((double) (endTime - startTime)) / spacing) + 1;
     }
 
     @Override
